Move ReviewAppBanner to PageBlock; soften styling
Renamespace the banner under PageBlock::ReviewAppBanner (matching the other
page_block ViewComponents) and update the i18n scope/render call to match.

Tweak the banner: drop the bold weight and lighten it (yellow-100 bg, smaller
text) so it's a quieter notice, and move the "email inbox" link to the end.
