+++ app/components/page_block/review_app_banner/component.en.yml
@@ -1,0 +1,9 @@
+---
+en:
+  components:
+    page_block:
+      review_app_banner:
+        disclaimer: not production, data is ephemeral
+        label: Review app
+        letter_opener_link: email inbox
+        pr_link: 'PR #%{number}'
