+++ app/components/page_block/review_app_banner/component_preview.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module PageBlock
+  module ReviewAppBanner
+    class ComponentPreview < ApplicationComponentPreview
+      def without_pr_link
+        render(PageBlock::ReviewAppBanner::Component.new(review_app: "1"))
+      end
+
+      def with_pr_link
+        render(PageBlock::ReviewAppBanner::Component.new(review_app: "1", pr_number: 1234))
+      end
+    end
+  end
+end
