+++ app/views/layouts/application.html.erb
@@ -48,7 +48,7 @@
   </head>
 
   <body id="<%= page_id %>" class="<%= body_class %>">
-    <%= render(ReviewAppBanner::Component.new(review_app: ENV["REVIEW_APP"], pr_number: ENV["REVIEW_APP_PR_NUMBER"])) %>
+    <%= render(PageBlock::ReviewAppBanner::Component.new(review_app: ENV["REVIEW_APP"], pr_number: ENV["REVIEW_APP_PR_NUMBER"])) %>
 
     <%# Cache navbar to speed stuff up, on per-user, per-page %>
     <% cache(["main_navbar4", page_id, current_user_or_unconfirmed_user, passive_organization]) do %>
